requirements.txt
# 深度学习框架
torch>=1.9.0
torchvision>=0.10.0

# 数据处理和科学计算
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# 图像处理
opencv-python>=4.5.0
Pillow>=8.3.0

# 数据可视化
matplotlib>=3.4.0
seaborn>=0.11.0

# 进度条和工具
tqdm>=4.62.0
requests>=2.25.0

# Jupyter支持（可选，用于交互式演示）
jupyter>=1.0.0
ipywidgets>=7.6.0

# 开发工具（可选）
black>=21.0.0
flake8>=3.9.0
pytest>=6.2.0
